search command
This commit adds a search command where you can search for a variety of
entities:

    fbarc search page blacklivesmatter
    fbarc search place minneapolis

Searching for user, event, group require your app be out of sandbox mode
(approved by FB).

Missing from this commit are unittests (I couldn't quite figure out what was
going on with the mocks). Also it seemed like paging of some kind would be
useful, and that there was support for that that could be leveraged?